Fix missing method and moved client
